This patch is not as big as it appears, it's mostly just changing some ifdefs and including the libs/libavcodec/mmx.h instead of the libs/libmythtv/mmx.h.
Jay Merrifield's mmx for x86-64 patch. This should enable mmx for
x86-64 systems everywhere except the in the linear_blend filter.
If someone wants to volunteer to rewrite that with the mmx.h
macros that would be great...

I also commented out the mm_support() in osdsurface.cpp, the problem
here appears to have been that dsputil.h was included as a C++ header.
It is actually a C header, but I've added the extern "C" {} construct
so it should work fine now.


git-svn-id: http://svn.mythtv.org/trunk@6351 4b91aca1-2a1f-0410-a73a-a78f2df169e8