Added --prefix to the configure script, for where to find MythTV. Make simlink to mythconfig.mak so that settings.pro can include it before it knows the prefix.
Included mythconfig.mak in settings.pro and copied over the
library and include stuff from MythTV's ./configure

There is some redundancy in the compile flags at the moment.
The .pro files in the varius plugins need to be adapted to this
change. But the important thing is that it should no longer be
necessary to modify settings.pro when using a non-pentium2
compatible CPU.


git-svn-id: http://svn.mythtv.org/trunk@6373 4b91aca1-2a1f-0410-a73a-a78f2df169e8