This is commit is to make John Pullan's life a little easier.
SignalMonitor is a base class for signal monitoring classes.
SignalMonitorValue is a helper class encapsulating a value such as
  "signal" or "signal_lock"
DTVSignalMonitor is an (incomplete) base class for digital signal monitors.
  pcHDTVSignalMonitor is a new class which detects the signal when
    using the v4l driver for the pcHDTV HD-2000/3000 cards.
  *DVBSignalMonitor is now a child class but is not updated in this commit.

The .cvsignore in libs/libmythtv/mpeg is just a copy of the one for
libs/libmythtv to avoid some cvs nagging.

None of these new classes is enabled in this commit, they are
there because John and I began working on the same code this week
and want to synchronize a bit.


git-svn-id: http://svn.mythtv.org/trunk@6498 4b91aca1-2a1f-0410-a73a-a78f2df169e8