This commit doesn't add any features.. These are the safe parts of the interactive signal monitoring patch, which I'm committing now so that John Pullan can get on with his analog signal monitoring code.
I've tested this with the pcHDTV HD-3000 with both v4l and DVB drivers.

In ChannelBase and Channel CheckSignalStrength() has been removed
and GetFd() has been added. The signal functionality is now
provided by the SignalMonitor classes, and GetFd() is used by
the pcHDTV and DVB signal monitoring classes. For the same
reasons DVBChannel no longer has a 'monitor' field, and has a
GetFd().

In DTVRecorder I've added QObject as an ancestor, this along with
ParseMGT and ParseVCT in HDTVRecorder() allows ATSCStreamData
to function without a pointer to the recorder via Qt signals.

I've moved the major and minor channel fields from MPEGStreamData
to ATSCStreamData, they never should have been in MPEGStreamData
which only cares about the program number. I've also added signals
for the various tables one might wish to know about, and I've added
Print slots for those tables so you can connect the two for
debugging purposes.

DVBSignalMonitor now inherits DTVSignalMonitor. It no longer has
the Status signal but it has the generic Lock signal which is all
we used the status for anyway. It also tests for whether things
like BER reporting are available on the card, and removes it from
the signal list if they aren't. Also like other SignalMonitor
classes it can be serialized to a QStringList, which is needed for
the interactive signal monitoring.

ScanWizard has been modified to use the Lock signal instead of
the Status signal, and to instanciate it's own DVBSignalMonitor
class since this is no longer in the dvbchannel class.

Finally, in TVRec I've removed the CheckSignalStrength() call
since this call no longer exists in Channel, but the functionality
will be back for more hardware types via the SignalMonitor soon.


git-svn-id: http://svn.mythtv.org/trunk@6506 4b91aca1-2a1f-0410-a73a-a78f2df169e8