From Joe Votour: make use of OpenGL Vsync a toggleable option.
The option to enable/disable OpenGL VSync is located in
Utilities/Setup -> Setup -> TV Settings -> Playback, on the
Hardware Decoding Settings page.

The big advantage of this patch is that binary packagers can
enable the OpenGL VSync code in their packages, rather than
having to leave it out due to instability on some people's
systems. OpenGL VSync defaults to off.


git-svn-id: http://svn.mythtv.org/trunk@6509 4b91aca1-2a1f-0410-a73a-a78f2df169e8