Audio sync adjustment patch from Mark Spieth.
Also a fix for a case where playback would never start due to there not being enough free buffers (because 0-1 is a REALLY big number in a long long)

Also a basic "native" DVD player.  It's not in it's final implementation, and still has quite a bit of work to go.  But it does work for main title playback. (yes menu support is planned)

You'll need to run configure again... If libdvdnav is detected the dvd playback code will be included.  Use --disable-dvd to turn this off regardless of libdvdnav.


git-svn-id: http://svn.mythtv.org/trunk@6552 4b91aca1-2a1f-0410-a73a-a78f2df169e8