Add a check at the top of AudioOutputBase::Reconfigure(). I don't think this will break anything else, and should solve the following problem:
    If you have an AudioOutput device that wants to stay locked on a given
frequency (eg. AudioOutputNull wants to always be at 44.1, no matter what
the input is at) then it will always set audio_samplerate to that fixed
frequency in OpenDevice().

    If Reconfigure() is called with input source at, say 22.05, everything
will work as expected; the resampler will kick in, and the output will come
out at 44.1.

    If you then switch to another source input that is at 44.1, the check at
the top of Reconfigure would return immediately (source is 44.1, locked
audio_samplerate is also 44.1), without turning off the resampler.

    So ... long story short, I've added a check to return immediately only
if the input and output parametes match AND need_resampler is not currently
set.


git-svn-id: http://svn.mythtv.org/trunk@6572 4b91aca1-2a1f-0410-a73a-a78f2df169e8