Nothing in this commit adds any user visible features to MythTV.
libmythtv.pro has been modified to let the makefile compile dvbtables and
dvbstreamdata, now that mpegstreamdata is synced with my local version.
hdtvrecorder has simply been updated to connect to the new mpegstreamdata
signals (described below), and atsctables.cpp has just had it's formatting
cleaned up.

mpegstreamdata now has a better division between the methods that apply
when you are using it to filter out a single program, and when you want
info on all the programs. To that end there are now two signals emitted
on each new PATs and most new PMTs, one UpdatePAT()/UpdatePMT() emits the
table actually in the source stream, while UpdatePATSingleProgram(),
UpdatePMTSingleProgram() emit tables processed to contain just the one
program we are interested in. UpdatePMTSingleProgram() is actually only
emitted when we see a PMT in the desired stream so unlike UpdatePMT(),
which is fired for each program in the stream, no pid is associated
with the signal.

atscstreamdata has had it's formatting cleanup up and also the TSID
is sent along with VCT tables instead of the PID. The TSID is
supposed to be unique, but the PID may not be... Both are available
from the VCT table itself, though getting the pid from a PSIP packet
is a bit baroque.

Finally, there have been some refactorings and new functionality added to
the signalmonitoring classes. The flags have been moved from
DTVSignalMonitor to SignalMonitor, and the "Signal Strength" signal
has been made optional, while all current implementations support
such a thing, it is no longer strictly needed, only "Signal Lock"
is required. In DTVSignalMonitor, I've added the ability to detect
PAT, PMT, MGT, and VCT tables. There are stubs for NIT and SDT tables
as well. DVBSignalMonitor and pcHDTVSignalMonitor have been modified
to start looking for the needed tables, if it is told either which
program number or minor and major channels it is looking for.
Service ID need's to be added for the DVBSignalMonitor, but I'm
just getting the ATSC stuff to work first.


git-svn-id: http://svn.mythtv.org/trunk@6602 4b91aca1-2a1f-0410-a73a-a78f2df169e8