This makes the ProgramInfo* parameter to the NVP constructor const, since we don't modify it. This allows us to get rid of the const removal cast in TVRec::GetScreenGrab().
This also adds a new setting "AudioNag" which if enabled, the default,
nags you about not being able to initialize the audio output in NVP.
We already have this nagging dialog, but it now gives you two new
options, "don't inform me again in this session" and "never inform me
again". In case you chose the latter and later regret it I've put this
setting into the General playback settings pane in globalsettings.cpp.


git-svn-id: http://svn.mythtv.org/trunk@6605 4b91aca1-2a1f-0410-a73a-a78f2df169e8