These changes are to support two improvements faster DTV channel changing and a more generic channel scanning class. There is a database update, the pidcache table is added.
The new methods in channelbase are:
  GetDevice() -- Returns tuner device, for use in siscan in place of GetCardNum()
  GetCachedPids()  -- Returns list of pids, and the tables they might contain
  SaveCachedPids() -- Saves list of pids and the tables they contain
  SetCachedATSCInfo()
  GetProgramNumber()/GetMajorChannel()/GetMinorChannel()

SetCachedATSCInfo() parses the channel name in Channel when dealing with
an ATSC channel, it allows you to use names without dashes or underscores.
The Get's get the parsed info. In DVBChannel the atscsrcid is used instead.
I plan on using that for the v4l HDTV drivers too, but not until I've
gotten the channel scanning to work for those drivers.

The cached pids functionality is not taken advantage of in this commit.
I'm using the signal monitor to save this as soon as it is safe, and the
signal monitor patch is not ready to be committed. The pidcache table
is simple and stable so I feel it is safe to commit it now.

There I've also moved the DVBTuning printing and parsing from DVBChannel
to DVBTuning. There is obviously some duplication in dvbtypes. I wrote
this before John Pullan created the parsing classes. But I'm not actually
increasing duplication, I'm just moving those duplicated parsers into the
same file which should make addressing the merge easier.


git-svn-id: http://svn.mythtv.org/trunk@6657 4b91aca1-2a1f-0410-a73a-a78f2df169e8