You must re-run ./configure (with the same options as normal).
This adds 3 options to ./configure
 --disable-v4l
 --disable-frontend
 --disable-backend

And, of course, the matching --enable options.

If you disable the frontend the backend can not create pixmaps for
a frontend for priviews, so it should be considered as something
useful only for porting the backend. Also if you do not compile
both the frontend and backend, mythtv-setup will not be compiled.
Disabling Video4Linux should be enough for compiling on MacOS X,
but on MacOS X the backend is disabled by default.

This touches a bunch of files, but it's mostly just for
USING_FRONTEND/BACKEND/V4L defines. The exception is that I added
a tv.cpp, because StateToString() is used by both tv_rec.cpp and
tv_play.cpp, and I didn't want to duplicate it in tv_rec.cpp and
tv_play.cpp nor did I want to make it static in tv.h.

The basic aim here is to make it possible for us to test frontend
only compilation so that Nigel doesn't have to do all the work on
keeping the MacOS X port working. And the secondary aim (i.e. why
do we have v4l and backend options) is to facilitate porting the
backend to MaxOS X.


git-svn-id: http://svn.mythtv.org/trunk@6672 4b91aca1-2a1f-0410-a73a-a78f2df169e8