This adds Tune() and TuneMultiplex() to the Channel class for channel scanning use.
This also more accurately sets the frequency from the frequency table
to the tuned to frequency. This means if you are using finetune, you
probably don't need it anymore. If you are using finetune you will have
to retune it.

This also allows you to optionally use the dtv_multiplex table for
additional tuning data.



git-svn-id: http://svn.mythtv.org/trunk@6705 4b91aca1-2a1f-0410-a73a-a78f2df169e8