This is an UGLY hack for a problem in CreateXvMCSurfaces().
reateXvMCSurfaces is supposed to allocate between 8 and 16 buffers,
allocating as many as possible, and reinitializing vbuffers as needed.
But it seems with the Unichrome chipset you need to call vbuffers.Init()
with the exact number of buffers you are capable of allocating BEFORE 
CreateXvMCSurfaces is called to allocate them.

Since the Unichrome chipset is the only chipset, I know of, which allows
more than eight XvMC buffers this is probably a general problem, exposed
simply by the unichrome chipset.

The hack is curtesy of Terry Barnaby, and should at least solve the
prebuffering pause problem with VLD until someone can look at why
CreateXvMCSurfaces isn't working proberly.

See #2 for a related problem.


git-svn-id: http://svn.mythtv.org/trunk@6728 4b91aca1-2a1f-0410-a73a-a78f2df169e8