Moved dvb_on_demand channel opening and closing from DVBRecorder to TVRec.
Added IsOpen() to ChannelBase(), to simplify determining when
we need to open or close the channel in tv_rec.

Created CloseChannel() method which handles closing the
channel iff we should be closing it.

Added a Close() call to Channel::SetFd() so that we don't leak
open file descriptors.

Put common channel initialization code from the different channel
types in TVRec::Init() to a InitChannel() method to reduce code
duplication.

Renamed CloseDVB() method in DVBChannel to Close() to match other
Channel classes. It should only be called in CloseChannel() if 
dvb_on_demand is not enabled, so the behavior does not change.

The only functional change is that you can start a DVBRecorder 
without reopening the DVBChannel when dvb_on_demand is enabled.
This is useful for signal monitoring because you tune to the 
channel before starting up the recorder and don't want the 
extra delay of having to re-tune to the service when you start
up the recorder.



git-svn-id: http://svn.mythtv.org/trunk@6856 4b91aca1-2a1f-0410-a73a-a78f2df169e8