Extended the use of the oldrecorded table to log what programs aren't recorded and why.  The new oldrecorded.duplicate field is used flag entries that should be considered for duplicate detection. Uninteresting entries are deleted automatically after 60 days.
When ProgramInfo objects are read using ProgramList::FromOldRecorded,
the ProgramInfo::duplicate member is set to the value of
oldrecorded.duplicate.  bjm updated previousbox to use this to color
code unrecorded programs appropriately.

When ProgramInfo objects are read using ProgramList::FromProgram, the
ProgramInfo::duplicate member is set if the program was previously
recorded.


git-svn-id: http://svn.mythtv.org/trunk@6896 4b91aca1-2a1f-0410-a73a-a78f2df169e8