Quick fix from David Engel for a problem I introduced in 6881. Basically if the audio changes in a stream in the commercial flagger then it will try to grab your audio device. This is bad.
So the quick fix is to add not create new audio outputs in
reinitaudio if using_null_videoout is true. Since the null
video out is only used in commercial flagging and transcoding
this serves as a good proxy for when we don't want audio.

The right fix would be to use add using_audio_out type variable
and check that, perhaps using the existing disableaudio (which
currently has a slightly different use). But that could introduce
bugs, and I have enough to track right now.



git-svn-id: http://svn.mythtv.org/trunk@6899 4b91aca1-2a1f-0410-a73a-a78f2df169e8