Another channel scanner commit
Changes SIScan to use EITHelper class to process EITs.
This SIScan change should be code neutral.

Also modified tv_rec.cpp to start up the SIScan after
DVBChannel creates an SIParser instance, the current
code has no chance of working.
TVRec is also added as a friend of DVBChannel so it
can check for the presence of the SIParser.

I'm still not getting EIT events, but this is probably
because I'm in ATSC Land.



git-svn-id: http://svn.mythtv.org/trunk@6922 4b91aca1-2a1f-0410-a73a-a78f2df169e8