Adds more debugging to GetBetterMplexID() on -v siparser
More importantantly this fixes a bug in channel scanning 
where the wrong mplexid was used when trying to insert a
channel into the DB.

This was caused when I tried to make the EIT parser work
again and used the DVBChannel's mplexid even if the
scanner had a better mplex id. The fix should have
been to use the scanner's mplex id, and only fall back
to the DVBChannel's mplexid if the scanner did not have
one. This is because when we tune to a channel in the
scanner an mplexid may not yet exist, so DVBChannel
does not have a valid mplexid.

Anyway, this should fix the problem.



git-svn-id: http://svn.mythtv.org/trunk@6951 4b91aca1-2a1f-0410-a73a-a78f2df169e8