Partial fix for the mpeg-ts problem.
I've been using this locally for a couple days now,
so it should be fairly stable.

In mpegts.c/mpegts.h & utils.c/avformat.h there are three changes:
  1/ When the streams change, packets in the old streams are flushed.
  2/ All codecs are closed when streams change.
  3/ Streams are changed all at once after scanning the whole PMT.

#2 should fix some of the audio problems during channel changing.
#1, #2, #3 together seem to have solved the problem of ffmpeg writing
past the end of the audio buffer.

In avformatdecoder, there have been some adjustments for #2 above, plus
there is much more debug info.

In videobuffers.cpp, I've increased the safety below each allocated
buffer, this allocator is used by VideoOutputNull.cpp. I'll reduce
this again once I figure out the remaining mpegts problem, this
prevents a segfault when generating previews for a video with
multiple stream changes in the first few seconds of a video.

This doesn't completely solve the problem however, ffmpeg is no longer
writing past the end of audio buffers, but it is still writing past the
end of the video buffers. Worse it looks like it is doing this on the
stack. This is in AvFormatDecoder::GetFrame() and it makes the
backtraces a bit "interesting".

BTW A workaround to the remaining problem is to use XvMC...



git-svn-id: http://svn.mythtv.org/trunk@6990 4b91aca1-2a1f-0410-a73a-a78f2df169e8