This removes the USE_OWN_SIPARSER define, since this feature appears to be working for everyone.
This adds the USE_SIPARSER define, which will allow us to disable
the SIParser completely in SIScan and just use the info from the 
DVBSignalMonitor, after some more testing in Europe.

If you comment out the USE_SIPARSER define now it will only work 
with ATSC streams (with either the DVB or V4L drivers).



git-svn-id: http://svn.mythtv.org/trunk@7008 4b91aca1-2a1f-0410-a73a-a78f2df169e8