Modifies SIScan to work with the DVBChannel from the Signal Monitor patch.
Also allows "Existing Transport Scan" to work with ATSC using DVB drivers,
but only on the second scan.

Removes verifyTransport() in siscan.cpp, whose function was already 
performed when a TransportScanItem was created from a frequency table.

This also fills in the implementation of
  void SIScan::UpdateSDTinDB(int, const ServiceDescriptionTable*, bool)
Which allows non-ATSC DVB driver users to try undefining USE_SIPARSER
and potentially get something that works.



git-svn-id: http://svn.mythtv.org/trunk@7014 4b91aca1-2a1f-0410-a73a-a78f2df169e8