Redo of changeset [7018].
Basically the difference is that DVBRecorder::SetDemuxFilters()
no longer puts the DVBRecorder in an error state if the PID's
are not set. The problem was that since DVBChannel no longer 
blocks until a good signal is seen, the recorder was more likely
to call SetDemuxFilters() before the recorder's PMT was set.



git-svn-id: http://svn.mythtv.org/trunk@7027 4b91aca1-2a1f-0410-a73a-a78f2df169e8