Closes #222.
This is Jasper's patch to force the non-broken TS format
for DVB recordings.

I'm not going to remove the ts->ps code yet. But I made it
harder to re-enable it by commenting out the config option.

We can remove the ts->ps code after 0.19, as we'll probably
get all the bug reports on TS playback problems we need 
after the release.



git-svn-id: http://svn.mythtv.org/trunk@7059 4b91aca1-2a1f-0410-a73a-a78f2df169e8