References #13.
This should go a long way toward addressing problem 1
in the list of tuning problems.

If a startchan is set we use it.
If it is not set we try to use a channel on the current input.
If there are no channels defined for the current input,
   we try other inputs.
If there are no channels defined at all for the card,
   we set the starting channel to '3', which should be
   fine for tunerless inputs (S-Video capture, etc).

I don't think the channel crawl for a start channel
is worth it since we won't have the getting the 
"Stuck in LiveTV" problem with the signal monitoring
feature.



git-svn-id: http://svn.mythtv.org/trunk@7063 4b91aca1-2a1f-0410-a73a-a78f2df169e8